+++ src/path.mli
@@ -24,4 +24,3 @@ val relative_all : t -> string list -> t
 val with_ext : t -> ext:string -> t
 val parent : t -> t option
 val is_root : t -> bool
-val asset : string -> t