+++ src/treeview_commands.ml
@@ -1,10 +1,10 @@
-let select_sandbox_item =
+let select_sandbox_item extension =
   let icon =
     `LightDark
-      Vscode.LightDarkIcon.
-        { light = `String (Path.asset "collection-light.svg" |> Path.to_string)
-        ; dark = `String (Path.asset "collection-dark.svg" |> Path.to_string)
-        }
+      (Extension_assets.light_dark_icon
+         ~extension
+         ~light:"collection-light.svg"
+         ~dark:"collection-dark.svg")
   in
   let label = `TreeItemLabel (Vscode.TreeItemLabel.create ~label:"Select a Sandbox" ()) in
   let item = Vscode.TreeItem.make_label ~label () in
@@ -16,13 +16,13 @@ let select_sandbox_item =
   item
 ;;
 
-let terminal_item =
+let terminal_item extension =
   let icon =
     `LightDark
-      Vscode.LightDarkIcon.
-        { light = `String (Path.asset "terminal-light.svg" |> Path.to_string)
-        ; dark = `String (Path.asset "terminal-dark.svg" |> Path.to_string)
-        }
+      (Extension_assets.light_dark_icon
+         ~extension
+         ~light:"terminal-light.svg"
+         ~dark:"terminal-dark.svg")
   in
   let label =
     `TreeItemLabel (Vscode.TreeItemLabel.create ~label:"Open a sandboxed terminal" ())
@@ -39,7 +39,7 @@ let terminal_item =
   item
 ;;
 
-let construct_item =
+let construct_item () =
   let icon = `ThemeIcon (Vscode.ThemeIcon.make ~id:"tools" ()) in
   let label =
     `TreeItemLabel
@@ -54,7 +54,7 @@ let construct_item =
   item
 ;;
 
-let jump_item =
+let jump_item () =
   let icon = `ThemeIcon (Vscode.ThemeIcon.make ~id:"fold-up" ()) in
   let label =
     `TreeItemLabel (Vscode.TreeItemLabel.create ~label:"Jump to a specific target" ())
@@ -66,7 +66,7 @@ let jump_item =
   item
 ;;
 
-let type_search_item =
+let type_search_item () =
   let icon = `ThemeIcon (Vscode.ThemeIcon.make ~id:"search-view-icon" ()) in
   let label =
     `TreeItemLabel
@@ -84,7 +84,7 @@ let type_search_item =
   item
 ;;
 
-let navigate_holes_item =
+let navigate_holes_item () =
   let icon = `ThemeIcon (Vscode.ThemeIcon.make ~id:"breakpoints-activate" ()) in
   let label =
     `TreeItemLabel (Vscode.TreeItemLabel.create ~label:"Navigate between typed holes" ())
@@ -101,25 +101,27 @@ let navigate_holes_item =
   item
 ;;
 
-let items =
-  [ select_sandbox_item
-  ; terminal_item
-  ; construct_item
-  ; jump_item
-  ; type_search_item
-  ; navigate_holes_item
+let items extension =
+  [ select_sandbox_item extension
+  ; terminal_item extension
+  ; construct_item ()
+  ; jump_item ()
+  ; type_search_item ()
+  ; navigate_holes_item ()
   ]
 ;;
 
 let getTreeItem ~element = `Value element
 
-let getChildren ?element () =
+let getChildren items ?element () =
   match element with
   | None -> `Value (Some items)
   | Some _ -> `Value (Some [])
 ;;
 
 let register extension =
+  let items = items extension in
+  let getChildren = getChildren items in
   let module TreeDataProvider = Vscode.TreeDataProvider.Make (Vscode.TreeItem) in
   let treeDataProvider = TreeDataProvider.create ~getTreeItem ~getChildren () in
   let disposable =